chore: rename np locale file
